Prevent spectators from interacting with the creative inventory

+++ src/main/java/cn/nukkit/entity/item/EntityMinecartTNT.java
@@ -144,7 +144,7 @@ public class EntityMinecartTNT extends EntityMinecartAbstract implements EntityE
     public void saveNBT() {
         super.saveNBT();
 
-        super.namedTag.putInt("TNTFuse", this.fuse);
+        super.namedTag.putByte("TNTFuse", this.fuse);
     }
 
     @Override
